% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob.R
\name{calc_test}
\alias{calc_test}
\title{Compute conditional probabilities of the genotypes given a sequence
of block markers}
\usage{
calc_test(m, rf)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Compute conditional probabilities of the genotypes given a sequence
of block markers
}
